Fix default tests, tidy up molecule
Begin some of the deduplication of molecule.yml available in Molecule
2.20
Fix the shared tests directory. It should be relative to the scenario's
"tests" directory and not to the scenario directory itself. Also, you
have to have at least an empty test_*.py file in the scenario's specific
tests directory before additional dirs are picked up. So now that is a
thing. Also, adding "*" would cause the test runner to try and pick up
the yaml files and others in that directory for testing. And that
doens't work really well.